fix calendar folder structure
