+++ packages/twenty-server/src/modules/calendar/commands/crons/google-calendar-sync.cron.command.ts
@@ -4,15 +4,14 @@ import { Command, CommandRunner } from 'nest-commander';
 
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/engine/integrations/message-queue/services/message-queue.service';
-import { GmailPartialSyncCronJob } from 'src/modules/messaging/jobs/crons/gmail-partial-sync.cron.job';
-import { fetchAllWorkspacesMessagesCronPattern } from 'src/modules/messaging/jobs/crons/patterns/fetch-all-workspaces-messages.cron.pattern';
+import { googleCalendarSyncCronPattern } from 'src/modules/calendar/commands/crons/patterns/google-calendar-sync.cron.pattern';
+import { GoogleCalendarSyncCronJob } from 'src/modules/calendar/jobs/crons/google-calendar-sync.cron.job';
 
 @Command({
-  name: 'cron:messaging:gmail-partial-sync',
-  description:
-    'Starts a cron job to sync existing connected account messages and store them in the cache',
+  name: 'cron:calendar:google-calendar-sync',
+  description: 'Starts a cron job to sync google calendar for all workspaces.',
 })
-export class GmailPartialSyncCronCommand extends CommandRunner {
+export class GoogleCalendarSyncCronJobCommand extends CommandRunner {
   constructor(
     @Inject(MessageQueue.cronQueue)
     private readonly messageQueueService: MessageQueueService,
@@ -22,10 +21,10 @@ export class GmailPartialSyncCronCommand extends CommandRunner {
 
   async run(): Promise<void> {
     await this.messageQueueService.addCron<undefined>(
-      GmailPartialSyncCronJob.name,
+      GoogleCalendarSyncCronJob.name,
       undefined,
       {
-        repeat: { pattern: fetchAllWorkspacesMessagesCronPattern },
+        repeat: { pattern: googleCalendarSyncCronPattern },
       },
     );
   }
