+++ packages/twenty-server/src/modules/calendar/commands/crons/patterns/google-calendar-sync.cron.pattern.ts
